SelectGroup logic to fragment
