+++ app/src/main/res/layout/fragment_select_group.xml
